#242 fix: 근거리통신_비로그인시 카카오리다이렉트 적용 위한 컴포넌트 분리 (재렌더링 반복 문제 발생 해결)
